pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = ""
version = "0.1.0"
description = "Nobel Laureat RAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-cpp-python; sys_platform != 'win32'",  # optional if using server-only
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
scalesys = ""
